+++ Access/Changes.md
@@ -2,6 +2,9 @@
 
 ## v1.4.6 (In Development)
 Supports changes in **Flow** v2.0
+### New Methods
+- **Status**
+  - Added `.isSuccess` and `.isFailure`
 
 ## v1.4.5 - 02.10.2022
 Supports changes in **Flow** v1.17